Read sequences from a fasta file.

read_fasta() reads all sequences contained in a fasta file into a tibble (see Value). See FASTA format for the requirements of input file.

read_fasta(file, per_line = FALSE, unalign = FALSE)

Arguments

file
string. Path to the fasta file.

per_line
logical scalar. Whether sequences keep in one line or might span multiple lines. Specify FALSE if not sure, refer to Details.

unalign
logical scalar. Whether unalign aligned sequences, i.e. remove '-', see Details.

Value

tibble

name: character. sequence header (without '>')

seq: character. sequence itself, I named it seq instead of sequence to save 62.5 means sequence)

Details

Previously, I implemented fasta as a named character (in bioinfor::read_fasta()). But I found it more and more inconvenient as I used it. Many times when you manipulate the character, the name got lost. Finally, I decided to reimplement it as a tibble. That why this package come into being.

Actually, file can be anything as long as it's recognized by readr::read_lines().

If per_line isn't TRUE, read_fasta() will check it by detecting '>' and count line number. This (implemented by stringr::str_detect()) takes almost the same time as reading file content (implemented by readr::read_lines()). So specify TRUE if you are sure to save time. But this checking doesn't waste any time when sequences indeed span multiple lines. Additionly, a file contains odd number of lines will cause pre_line=TURE to be ignored

For unalign, when deal with large aligned file with linebreak containing many '-'s, implement unalign (stringr::str_replace_all() in read_fasta()) is about 15 read_fasta() returns.

FASTA format

The minimum requirement is that it can only contain sequences (can use any character except for '>', at least at the begining) and their headers (begin with '>', one line). Comments are not supported.

Basically, c('>na', 'me', 'ATCG') (name span two lines) and c('>name', '>ATCG') (sequence begins with '>') are most common (and maybe the only) errors. Expect that, almost anything is acceptable (though your file may not be recongnized by other program):

name contains '>', see crazy example 1.

sequence contains '>' (can't be at the begining), see crazy example 2.

empty name, see crazy example 3.

empty sequence, see crazy example 4. But it must occupy an empty line if you want to scape check for pre_line (refer to details).

very long name, see crazy example 5. It can contain thousands of thousands characters, so long as you have enough memory (and disk if you are really mad).

don't you think the above is enough? We can't make friends anymore. (Never tell me you have tried crazy example 6)

In most situations, you won't have a file like that, unless you deliberately create one.

Test files

test1.fasta big file with linebreak.

test12.fasta same file without.

test2.fasta big aligned file with linebreak.

Examples

system.file('extdata', 'example.fasta', package = 'biozhuoer') %>% read_fasta();

#> # A tibble: 62 x 2
#>    name                    seq                                                  
#>    <chr>                   <chr>                                                
#>  1 Acanthisitta_chloris_J… MVKLGSNLNDKNNKQPSNEDGFQTVPLITPLEVNHLQFPAPEKVIVKTRTEY…
#>  2 Amazona_aestiva_protein MLAQGKGPQAQCSSQCSSHDASQSHCPPQLPASAGPRHHPAPPHMAPGRRQG…
#>  3 Anas_platyrhynchos_Jar… MVKLGSNLNDKNSKPPSGEDGFQTVPLITPLEVNHLQFPAPEKVIVKTRTEY…
#>  4 Anser_cygnoides_protein MCEVSLGVACCVARNLGVKPRKPIRQGKVSQTGQKSSSHYPCDNDRLAGFRS…
#>  5 Antrostomus_carolinens… QVIVKTRTEYQPDQKNKGKLRVPKIAEFTVSFTDGVTERLKVTILISLALAF…
#>  6 Apaloderma_vittatum_Ja… MVKLGSNLSDKSSKQPPNEDGFQTVPLITPLEVNHLQFPAPEKVIVKTRTEY…
#>  7 Aptenodytes_forsteri_J… MVKLGSNLNDKNNKQPSNEDGFQTVPLITPLEVNHLQFPAPEKVIVKTRTEY…
#>  8 Apteryx_australis_prot… VTILISLALAFLACIVFLVVYKAFTYDHSCPDGFVYKHKRCIPASLDAYYSA…
#>  9 Aquila_chrysaetos_prot… MVKLGSNLNDKNSKQPSNEDGFQTVPLITPLEVNHLQFPAPEKVIVKTRTEY…
#> 10 Balearica_regulorum_Ja… MVKLGSNLNDKNNKQPSNEDGFQTVPLITPLEVNHLQFPAPEKVIVKTRTEY…
#> # ... with 52 more rows

system.file('extdata', 'aligned-multiline.fasta', package = 'biozhuoer') %>% read_fasta();

#> # A tibble: 61 x 2
#>    name                    seq                                                  
#>    <chr>                   <chr>                                                
#>  1 Acanthisitta_chloris_J… -----------------------------------------------AG---…
#>  2 Amazona_aestiva_protein MRFWGGGWTLRXT-SEMGNXS----------VQRMRE--------GPGG---…
#>  3 Anas_platyrhynchos_Jar… ----------------------------------------------RAG---…
#>  4 Anser_cygnoides_protein ----------------------------------------------------…
#>  5 Antrostomus_carolinens… -----------------------------------------------AG---…
#>  6 Apaloderma_vittatum_Ja… ----------------------------------------------RAG---…
#>  7 Aptenodytes_forsteri_J… -----------------------------------------------AG---…
#>  8 Apteryx_australis_prot… ----------------------------------------------------…
#>  9 Aquila_chrysaetos_prot… ---------------------------------------------MFSGCIQ…
#> 10 Balearica_regulorum_Ja… -----------------------------------------------AG---…
#> # ... with 51 more rows

system.file('extdata', 'aligned-multiline.fasta', package = 'biozhuoer') %>% read_fasta(unalign = TRUE);

#> # A tibble: 61 x 2
#>    name                    seq                                                  
#>    <chr>                   <chr>                                                
#>  1 Acanthisitta_chloris_J… AGVLALWALTTHGMYIQDFWRTWLRGLRFFLAVGIFFCVVALVAFCTFLALA…
#>  2 Amazona_aestiva_protein MRFWGGGWTLRXTSEMGNXSVQRMREGPGGVLALWALITHVMYVQDYWRTWL…
#>  3 Anas_platyrhynchos_Jar… RAGVLALWALITHVMYVQDYWRTWLKGLRFFLFIGILFSALSVVGICTFLVL…
#>  4 Anser_cygnoides_protein MYVQDYWRTWLKGLRFFLFIGILFSALSVVGFCTFLVLAITKHQSLTDPRSY…
#>  5 Antrostomus_carolinens… AGVLALWALITHVMYVQDYWRTWLKGLRFFLFIGILFSALSVVGFCTFLVLA…
#>  6 Apaloderma_vittatum_Ja… RAGVLALWALTTHVMYVQDYWRTWLKGLRFFLFIGILFSALSAVGFCAFLVL…
#>  7 Aptenodytes_forsteri_J… AGVLALWALITHVMYVQDYWRTWLKGLRFFLFIGILFSALSVVGFCTFLVLA…
#>  8 Apteryx_australis_prot… MHLSFPLESDYWRTWLKGLRFFLFIGILFSALSVVGFCTFLVLAITKHQSLT…
#>  9 Aquila_chrysaetos_prot… MFSGCIQPFPEWLLMFLCVLALWALITHVMYVQDYWRTWLKGLRFFLFIGIL…
#> 10 Balearica_regulorum_Ja… AGVLALWALITHVMYVQDYWRTWLKGLRFFLFIGILFSALSVVGFCTFLVLA…
#> # ... with 51 more rows

# crazy examples
read_fasta('>na>me\nATCG');

#> # A tibble: 1 x 2
#>   name  seq  
#>   <chr> <chr>
#> 1 na>me ATCG 

read_fasta('>name\nAT>CG');

#> # A tibble: 1 x 2
#>   name  seq  
#>   <chr> <chr>
#> 1 name  AT>CG

read_fasta('>\nATCG');

#> # A tibble: 1 x 2
#>   name  seq  
#>   <chr> <chr>
#> 1 ""    ATCG 

read_fasta('>name\n')

#> # A tibble: 1 x 2
#>   name  seq  
#>   <chr> <chr>
#> 1 name  ""   

read_fasta(paste0(c('>', rep('x', 10000), '\nATCG'), collapse = ''));

#> # A tibble: 1 x 2
#>   name                                                                     seq  
#>   <chr>                                                                    <chr>
#> 1 xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx… ATCG 

tempfile() %T>% readr::write_file('>name', .)  %T>% readr::read_file() %>% read_fasta;

#> # A tibble: 1 x 2
#>   name  seq  
#>   <chr> <chr>
#> 1 name  ""